ceps_client

Struct SDK

pub struct SDK { /* private fields */ }

Re-export of the underlying SDK type for advanced callers.

Implementations

impl SDK

pub async fn get_account( &self, account_identifier: Option<AccountIdentifier>, account_identifier_as_string: Option<String>, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetAccountResult>, SdkError>

👎Deprecated:
prefer ‘get_entity’

Retrieves account information based on the provided options.

Arguments

account_identifier - An optional AccountIdentifier for specifying the account identifier.

account_identifier_as_string - An optional string representing the account identifier.

maybe_block_identifier - An optional BlockIdentifierInput for specifying a block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a SuccessResponse<_GetAccountResult> or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_auction_info( &self, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetAuctionInfoResult>, SdkError>

Retrieves auction information based on the provided options.

Arguments

maybe_block_identifier - An optional BlockIdentifierInput for specifying a block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a _GetAuctionInfoResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_balance( &self, purse_uref: GetBalanceInput, state_root_hash: Option<impl ToDigest>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetBalanceResult>, SdkError>

Retrieves balance information based on the provided options.

Arguments

state_root_hash - The state root hash to query for balance information.

purse_uref - The purse uref specifying the purse for which to retrieve the balance.

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a GetBalanceResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_block( &self, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetBlockResult>, SdkError>

Retrieves block information using the provided options.

Arguments

maybe_block_identifier - An optional BlockIdentifierInput specifying the block identifier.

verbosity - An optional Verbosity level for the retrieval.

rpc_address - An optional rpc address to target for retrieval.

Returns

A Result containing either a _GetBlockResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_block_transfers( &self, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetBlockTransfersResult>, SdkError>

Retrieves block transfers information based on the provided options.

Arguments

maybe_block_identifier - An optional BlockIdentifierInput specifying the block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a _GetBlockTransfersResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

Implementations for the SDK struct.

pub async fn get_chainspec( &self, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetChainspecResult>, Error>

Asynchronously retrieves the chainspec.

Arguments

verbosity - An optional Verbosity parameter.

rpc_address - An optional rpc address as a string.

Returns

A Result containing either a _GetChainspecResult or a SdkError in case of an error.

impl SDK

pub async fn get_deploy( &self, deploy_hash: DeployHash, finalized_approvals: Option<bool>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetDeployResult>, Error>

👎Deprecated:
prefer ‘get_transaction’

Retrieves deploy information based on the provided options.

Arguments

deploy_hash - The deploy hash.

finalized_approvals - An optional boolean indicating finalized approvals.

verbosity - An optional verbosity level.

rpc_address - An optional rpc address.

Returns

A Result containing either a _GetDeployResult or an error.

impl SDK

pub async fn get_dictionary_item( &self, dictionary_item_input: DictionaryItemInput, state_root_hash: Option<impl ToDigest>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetDictionaryItemResult>, SdkError>

Retrieves dictionary item information based on the provided options.

Arguments

state_root_hash - A ToDigest implementation for specifying the state root hash.

dictionary_item - A DictionaryItemInput enum specifying the dictionary item to retrieve.

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a _GetDictionaryItemResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_entity( &self, entity_identifier: Option<EntityIdentifier>, entity_identifier_as_string: Option<String>, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetAddressableEntityResult>, SdkError>

Retrieves account information based on the provided options.

Arguments

entity_identifier - An optional EntityIdentifier for specifying the entity identifier.

entity_identifier_as_string - An optional string representing the entity identifier.

maybe_block_identifier - An optional BlockIdentifierInput for specifying a block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a SuccessResponse<_GetAddressableEntityResult> or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_era_info( &self, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetEraInfoResult>, SdkError>

👎Deprecated:
prefer ‘get_era_summary’ as it doesn’t require a switch block

impl SDK

pub async fn get_era_summary( &self, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetEraSummaryResult>, SdkError>

Retrieves era summary information based on the provided options.

Arguments

maybe_block_identifier - An optional BlockIdentifierInput for specifying a block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a _GetEraSummaryResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_node_status( &self, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetNodeStatusResult>, Error>

Retrieves node status information based on the provided options.

Arguments

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a _GetNodeStatusResult or an Error in case of an error.

Errors

Returns an Error if there is an error during the retrieval process.

impl SDK

pub async fn get_peers( &self, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetPeersResult>, Error>

Retrieves peers.

Arguments

verbosity - Optional verbosity level.

rpc_address - Optional rpc address.

Returns

A Result containing SuccessResponse with _GetPeersResult or an Error if an error occurs.

impl SDK

pub async fn get_reward( &self, validator: PublicKey, maybe_delegator: Option<PublicKey>, maybe_era_identifier: Option<EraIdentifier>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetRewardResult>, SdkError>

Retrieves reward information (info_get_reward).

Arguments

validator - Validator public key.

maybe_delegator - Optional delegator public key (validator reward when None).

maybe_era_identifier - Optional era or block identifier (None = last finalized era).

verbosity - Optional verbosity.

rpc_address - Optional RPC URL.

pub async fn get_reward_as_string( &self, validator: &str, maybe_delegator: Option<&str>, maybe_era_id: Option<&str>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetRewardResult>, SdkError>

CLI-string convenience wrapper around info_get_reward.

impl SDK

pub async fn get_state_root_hash( &self, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetStateRootHashResult>, SdkError>

Retrieves state root hash information based on the provided options.

Arguments

maybe_block_identifier - An optional BlockIdentifierInput for specifying a block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a _GetStateRootHashResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn get_transaction( &self, transaction_hash: TransactionHash, finalized_approvals: Option<bool>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetTransactionResult>, Error>

Retrieves transaction information based on the provided options.

Arguments

transaction_hash - The transaction hash.

finalized_approvals - An optional boolean indicating finalized approvals.

verbosity - An optional verbosity level.

rpc_address - An optional rpc address.

Returns

A Result containing either a _GetTransactionResult or an error.

impl SDK

pub async fn get_validator_changes( &self, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetValidatorChangesResult>, Error>

Retrieves validator changes based on the provided options.

Arguments

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a _GetValidatorChangesResult or an Error in case of an error.

Errors

Returns an Error if there is an error during the retrieval process.

impl SDK

pub async fn list_rpcs( &self, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<ListRpcsResult>, Error>

Lists available RPCs based on the provided options.

Arguments

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a _ListRpcsResult or an Error in case of an error.

Errors

Returns an Error if there is an error during the listing process.

impl SDK

pub async fn put_transaction( &self, transaction: Transaction, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<PutTransactionResult>, Error>

Puts a transaction based on the provided options.

Arguments

transaction - The Transaction object to be sent.

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a _PutTransactionResult or an Error in case of an error.

Errors

Returns an Error if there is an error during the transaction process.

impl SDK

pub async fn query_balance( &self, maybe_global_state_identifier: Option<GlobalStateIdentifier>, purse_identifier_as_string: Option<String>, purse_identifier: Option<PurseIdentifier>, state_root_hash: Option<String>, maybe_block_id: Option<String>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<QueryBalanceResult>, SdkError>

Retrieves balance information based on the provided options.

Arguments

maybe_global_state_identifier - An optional GlobalStateIdentifier for specifying global state.

purse_identifier_as_string - An optional string representing a purse identifier.

purse_identifier - An optional PurseIdentifier.

state_root_hash - An optional string representing a state root hash.

maybe_block_id - An optional string representing a block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a SuccessResponse<_QueryBalanceResult> or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn query_balance_details( &self, maybe_global_state_identifier: Option<GlobalStateIdentifier>, purse_identifier_as_string: Option<String>, purse_identifier: Option<PurseIdentifier>, state_root_hash: Option<String>, maybe_block_id: Option<String>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<QueryBalanceDetailsResult>, SdkError>

Retrieves balance information details based on the provided options.

Arguments

maybe_global_state_identifier - An optional GlobalStateIdentifier for specifying global state.

purse_identifier_as_string - An optional string representing a purse identifier.

purse_identifier - An optional PurseIdentifier.

state_root_hash - An optional string representing a state root hash.

maybe_block_id - An optional string representing a block identifier.

verbosity - An optional Verbosity level for controlling the output verbosity.

rpc_address - An optional string specifying the rpc address to use for the request.

Returns

A Result containing either a SuccessResponse<_QueryBalanceDetailsResult> or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the retrieval process.

impl SDK

pub async fn query_global_state( &self, query_params: QueryGlobalStateParams, ) -> Result<SuccessResponse<QueryGlobalStateResult>, SdkError>

Retrieves global state information based on the provided parameters.

Arguments

query_params - A QueryGlobalStateParams struct containing query parameters.

Returns

A Result containing either a SuccessResponse<_QueryGlobalStateResult> or a SdkError in case of an error.

impl SDK

pub async fn speculative_exec( &self, transaction: Transaction, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<SpeculativeExecTxnResult>, SdkError>

Perform speculative execution.

Arguments

transaction - The transaction to execute.

verbosity - The verbosity level for logging.

rpc_address - The address of the node to connect to.

Returns

A Result containing the result of _SpeculativeExecTxnResult or a SdkError in case of an error.

impl SDK

pub async fn speculative_transaction( &self, builder_params: TransactionBuilderParams, transaction_params: TransactionStrParams, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<SpeculativeExecTxnResult>, SdkError>

This function allows executing a transaction speculatively.

Arguments

builder_params - Transaction Builder parameters.

transaction_params - Transactionment parameters for the transaction.

verbosity - Optional verbosity level.

rpc_address - Optional rpc address.

Returns

A Result containing either a SuccessResponse<_SpeculativeExecTxnResult> or a SdkError in case of an error.

impl SDK

pub async fn speculative_transfer_transaction( &self, maybe_source: Option<URef>, target_account: &str, amount: &str, transaction_params: TransactionStrParams, maybe_id: Option<String>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<SpeculativeExecTxnResult>, SdkError>

Perform a speculative transfer transaction.

Arguments

maybe_source - Optional transfer source uref.

amount - The amount to transfer.

target_account - The target account.

transaction_params - The transactionment parameters.

maybe_id - An optional transfer ID (defaults to a random number).

verbosity - The verbosity level for logging (optional).

rpc_address - The address of the node to connect to (optional).

Returns

A Result containing the result a SuccessResponse<_SpeculativeExecTxnResult> or a SdkError in case of an error.

impl SDK

pub async fn transaction( &self, builder_params: TransactionBuilderParams, transaction_params: TransactionStrParams, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<PutTransactionResult>, SdkError>

Perform a transaction operation.

Arguments

builder_params - Transaction Builder parameters.

transaction_params - Transaction parameters.

verbosity - An optional verbosity level.

rpc_address - An optional rpc address.

Returns

A result containing a SuccessResponse<_PutTransactionResult> or an SdkError.

impl SDK

pub async fn transfer_transaction( &self, maybe_source: Option<URef>, target_account: &str, amount: &str, transaction_params: TransactionStrParams, maybe_id: Option<String>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<PutTransactionResult>, SdkError>

Perform a transfer transaction of funds.

Arguments

maybe_source - Optional transfer source uref.

target_account - The target account.

amount - The amount to transfer.

transaction_params - The transaction parameters.

maybe_id - An optional transfer ID (defaults to a random number).

verbosity - The verbosity level for logging (optional).

rpc_address - The address of the node to connect to (optional).

Returns

A Result containing the result a SuccessResponse<_PutTransactionResult> of the transfer or a SdkError in case of an error.

impl SDK

pub fn make_transaction( &self, builder_params: TransactionBuilderParams, transaction_params: TransactionStrParams, ) -> Result<Transaction, Box<SdkError>>

Creates a transaction using the provided parameters.

Arguments

builder_params - Transaction Builder parameters.

transaction_params - The transaction parameters.

Returns

A Result containing the created Transaction or a SdkError in case of an error.

impl SDK

pub fn make_transfer_transaction( &self, maybe_source: Option<URef>, target: &str, amount: &str, transaction_params: TransactionStrParams, maybe_id: Option<String>, ) -> Result<Transaction, Box<SdkError>>

Creates a transfer transaction with the provided parameters.

Arguments

maybe_source - Optional transfer source uref.

amount - The transfer amount.

target - The target account.

transaction_params - The transaction parameters.

maybe_id - Optional transfer identifier.

Returns

A Result containing the created Transaction or a SdkError in case of an error.

impl SDK

pub fn sign_transaction( &self, transaction: Transaction, secret_key: &str, ) -> Transaction

Signs a transaction using the provided secret key.

Arguments

transaction - The transaction to sign.

secret_key - The secret key for signing.

Returns

The signed Transaction.

impl SDK

pub fn watch_deploy( &self, events_url: &str, timeout_duration: Option<u64>, ) -> Watcher

👎Deprecated:
prefer ‘watch_transaction’

Creates a new Watcher instance to watch deploys. Legacy alias

Arguments

events_url - The URL to monitor for transaction events.

timeout_duration - An optional timeout duration in seconds.

Returns

A Watcher instance.

pub fn watch_transaction( &self, events_url: &str, timeout_duration: Option<u64>, ) -> Watcher

Creates a new Watcher instance to watch deploys.

Arguments

events_url - The URL to monitor for transaction events.

timeout_duration - An optional timeout duration in seconds.

Returns

A Watcher instance.

pub async fn wait_deploy( &self, events_url: &str, deploy_hash: &str, timeout_duration: Option<u64>, ) -> Result<EventParseResult, String>

👎Deprecated:
prefer ‘wait_transaction’ with transaction

Waits for a deploy event to be processed asynchronously. Legacy alias

Arguments

events_url - The URL to monitor for transaction events.

deploy_hash - The deploy hash to wait for.

timeout_duration - An optional timeout duration in milliseconds.

Returns

A Result containing either the processed EventParseResult or an error message.

pub async fn wait_transaction( &self, events_url: &str, target_hash: &str, timeout_duration: Option<u64>, ) -> Result<EventParseResult, String>

Alias for wait_deploy Waits for a deploy event to be processed asynchronously.

Arguments

events_url - The URL to monitor for transaction events.

target_hash - The transaction hash to wait for.

timeout_duration - An optional timeout duration in milliseconds.

Returns

A Result containing either the processed EventParseResult or an error message

impl SDK

pub fn SSE_client(&self, events_url: &str) -> SSEClient

Build an SSEClient for events_url (e.g. http://node:9999/events).

pub async fn CES_parser( &self, contract_hashes: &[String], state_root_hash: Option<&str>, rpc_address: Option<String>, ) -> Result<CESParser, String>

Fetch CES schemas for contract_hashes and return a ready [super::CESParser].

impl SDK

A set of functions for working with smart contract entry points. Alias of sdk.transaction

pub async fn call_entrypoint( &self, builder_params: TransactionBuilderParams, transaction_params: TransactionStrParams, rpc_address: Option<String>, runtime_v2: Option<bool>, ) -> Result<SuccessResponse<PutTransactionResult>, SdkError>

Calls a smart contract entry point and returns the put result.

When runtime_v2 is omitted, the builder runtime is left unchanged (entity/package builders default to VmCasperV1). Pass true for VmCasperV2 or false for VmCasperV1.

Arguments

builder_params - Transaction Builder parameters.

transaction_params - Transaction parameters.

rpc_address - An optional rpc address to send the request to.

runtime_v2 - None → keep builder; true → V2; false → V1.

Returns

A Result containing either a _PutTransactionResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the call.

impl SDK

A set of functions for installing smart contracts on the blockchain. Alias of sdk.transaction

pub async fn install( &self, transaction_params: TransactionStrParams, transaction_bytes: Bytes, rpc_address: Option<String>, runtime_v2: Option<bool>, ) -> Result<SuccessResponse<PutTransactionResult>, SdkError>

Installs session wasm (is_install_upgrade) and returns the put result.

Session runtime defaults to VmCasperV2 (runtime_v2 omitted or true). Pass runtime_v2: false for classic VmCasperV1 wasm (HELLO, CEP-78, …).

Arguments

transaction_params - Transaction parameters.

transaction_bytes - Transaction Bytes to install

rpc_address - An optional rpc address to send the request to.

runtime_v2 - None/true → VmCasperV2; false → VmCasperV1.

Returns

A Result containing either a _PutTransactionResult or a SdkError in case of an error.

Errors

Returns a SdkError if there is an error during the installation.

impl SDK

Alias of sdk.get_dictionary_item

pub async fn query_contract_dict( &self, dictionary_item: DictionaryItemInput, state_root_hash: Option<impl ToDigest>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<GetDictionaryItemResult>, SdkError>

Query a contract dictionary item.

Arguments

state_root_hash - State root hash.

dictionary_item - Dictionary item input.

verbosity - Optional verbosity level.

rpc_address - Optional rpc address.

Returns

A Result containing either a SuccessResponse<_GetDictionaryItemResult> or a SdkError in case of an error.

impl SDK

Alias of sdk.query_global_state

pub async fn query_contract_key( &self, entity_identifier: Option<EntityIdentifier>, entity_identifier_as_string: Option<String>, path: PathIdentifierInput, maybe_block_identifier: Option<BlockIdentifierInput>, verbosity: Option<Verbosity>, rpc_address: Option<String>, ) -> Result<SuccessResponse<QueryGlobalStateResult>, SdkError>

Query a contract key.

Arguments

query_params - Query global state parameters.

Returns

A Result containing either a SuccessResponse<_GetAddressableEntityResult> or a SdkError in case of an error.

impl SDK

pub fn new( rpc_address: Option<String>, node_address: Option<String>, verbosity: Option<Verbosity>, ) -> SDK

pub fn get_rpc_address(&self, rpc_address: Option<String>) -> String

pub fn set_rpc_address( &mut self, rpc_address: Option<String>, ) -> Result<(), String>

pub fn get_node_address(&self, node_address: Option<String>) -> String

pub fn set_node_address( &mut self, node_address: Option<String>, ) -> Result<(), String>

pub fn get_verbosity(&self, verbosity: Option<Verbosity>) -> Verbosity

pub fn set_verbosity( &mut self, verbosity: Option<Verbosity>, ) -> Result<(), String>